+++ packages/mdc-drawer/index.js
@@ -14,4 +14,7 @@
  * limitations under the License.
  */
 
+import * as util from './util';
 export {MDCTemporaryDrawer, MDCTemporaryDrawerFoundation} from './temporary';
+export {MDCPersistentDrawer, MDCPersistentDrawerFoundation} from './persistent';
+export {util};