+++ packages/mdc-radio/index.js
@@ -42,6 +42,14 @@ export class MDCRadio extends MDCComponent {
     this.foundation_.setDisabled(disabled);
   }
 
+  get value() {
+    return this.foundation_.getValue();
+  }
+
+  set value(value) {
+    this.foundation_.setValue(value);
+  }
+
   get ripple() {
     return this.ripple_;
   }